chore: change directory structure
