+++ src/components/MultiplePicker/interface.ts
@@ -1,6 +1,6 @@
 import { Rule } from 'rc-field-form/es/interface';
 import React from 'react';
-import { IAliasProps } from '../../DynamicForm';
+import { IAliasProps } from '../DynamicForm';
 import { ErrorValueProps } from '../../PropsType';
 
 export interface IDataItem {
