+++ src/components/NomarSelect/interface.ts
@@ -1,6 +1,6 @@
 import { Rule } from 'rc-field-form/es/interface';
 import { PickerData } from 'antd-mobile/lib/picker/PropsType';
-import { IAliasProps } from '../../DynamicForm';
+import { IAliasProps } from '../DynamicForm';
 import { PickerPropsType } from 'antd-mobile/es/picker/PropsType';
 import { ErrorValueProps } from '../../PropsType';
