+++ src/components/index.tsx
@@ -1,5 +1,6 @@
 export { default as AddressPicker } from './AddressPicker';
 export { default as CoverRadio } from './CoverRadio';
+export { default as DynamicForm } from './DynamicForm';
 export { default as ExtraInput } from './ExtraInput';
 export { default as InputItemAA } from './InputItemaaa';
 export { default as MultiplePicker } from './MultiplePicker';
@@ -19,3 +20,4 @@ export { default as OnlyReadInput } from './OnlyReadInput';
 export { default as RangeDatePicker } from './RangeDatePicker';
 export { default as TextItem } from './TextItem';
 export { default as InputItem } from './InputItem';
+export { default as Field } from './Field';
