+++ src/index.ts
@@ -1,14 +1,12 @@
-import DynamicForm from './DynamicForm';
-import Field from './components/Field';
-import Form from 'rc-field-form';
+import DynamicForm, { useForm } from './components/DynamicForm';
 import './index.less';
 
-export * from './DynamicForm';
 export * from './components';
 export * from './utils';
-export * from './components/Field';
-export * from 'rc-field-form/es/interface';
+export * from './PropsType';
 
-export { Field, Form, DynamicForm };
+export { useForm };
+// rename
+export { DynamicForm as Form };
 
 export default DynamicForm;
