+++ typings.d.ts
@@ -1,3 +1,4 @@
+declare module '*.png';
 declare module '*.css';
 declare module '*.less';
 declare module '@bang88/china-city-data';
